pyproject.toml
[tool.poetry]
name = "repl_python3_Cheems-Full"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
flask = "^2.0.0"
praw = "^7.2.0"
pymediawiki = "^0.7.0"
"akinator.py" = "^4.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
